DEBUG for FetchHTTP instance creation, initialTasks(), cleanupHttp().
Debug on FetchHTTP to help understand why the 'http_cookies' database is not
being closed. This issue is discussed in more detail at:
https://github.com/DIA-NZ/webcurator/issues/104.

This particular branch prints debug messages and stack traces to provide
insight as to when a FetchHTTP crawler is created and when initialTasks()
and cleanupHttp() are being called, as well as the state of the cookieDb
instance.

Note that a version of the debug jar has been checked in to the branch
in the release_archive folder. It can be installed using:
mvn install:install-file   -Dfile=<project-location>/release_archive/heritrix-1.14.3-webcuratortool-2.0.1-SNAPSHOT.jar   -DpomFile=<project-location>/release_archive/heritrix-1.14.3-webcuratortool-2.0.1-SNAPSHOT.pom