Adjustments for conflicts with openwayback-core:2.4.0.
pom.xml, project.xml:
- Version has been changed to 1.14.2-webcuratortool-2.0.2.
- The 1.14.2-webcuratortool-2.0.2 built version has been checked into release_archive.

If a downstream dependency is using org.netpreserve.openwayback:openwayback-core:2.4.0, there will be conflicts
with some of the classes in this version of heritrix. What usually happens is a class in this project will
call another class in the project, but that version of the class will come from openwayback-core, and the
openwayback-core version of the class won't have methods that the version in this project has.

In order to resolve the conflicts certain classes have been adjusted so as to work with openwayback-core classes
(when those conflicts have been encountered).

ArchiveUtils.java:
- Moved classnameBasedUID to new class ArchiveUtilsClassnameBasedUIDOnly.java.
- classnameBasedUID does not exist on openwayback-core based ArchiveUtils.

FetchHTTP.java, CrawlController.java, BdbFrontier.java, BdbMultipleWorkQueues.java, BdbWorkQueue.java,
FPUriUniqFilter.java, FTPException.java:
- Use ArchiveUtilsClassnameBasedUIDOnly.classnameBasedUID instead of ArchiveUtils.classnameBasedUID.

HeritrixSSLProtocolSocketFactory.java:
- Constructor call to ConfigurableX509TrustManager uses no parameters (which then uses DEFAULT, but that has
  changed from a String to TrustLevel).